Kernel/Graphics: Move Intel graphics related code to a separate folder
