+++ Kernel/Graphics/Intel/NativeGraphicsAdapter.cpp
@@ -7,7 +7,7 @@
 #include <Kernel/Graphics/Console/ContiguousFramebufferConsole.h>
 #include <Kernel/Graphics/Definitions.h>
 #include <Kernel/Graphics/GraphicsManagement.h>
-#include <Kernel/Graphics/IntelNativeGraphicsAdapter.h>
+#include <Kernel/Graphics/Intel/NativeGraphicsAdapter.h>
 #include <Kernel/IO.h>
 #include <Kernel/PhysicalAddress.h>
