+++ Kernel/Graphics/Intel/NativeGraphicsAdapter.h
